requirements.txt
flask 
groq
python-dotenv